Add Factory-like class for S calculators.
